# services/quiz/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API квизов, вопросов, ответов и пользователей quiz-service.\nПубличные ответы не содержат is_correct, а пользователи — хешей паролей.",
        "title": "QuizGo CRUD API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/answers": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только автору квиза. Если is_correct не передан, используется false. Тело запроса должно содержать один JSON-объект без неизвестных полей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Создать ответ",
                "parameters": [
                    {
                        "description": "Новый ответ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quizhandler.CreateAnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ответ создан; тело ответа пустое"
                    },
                    "400": {
                        "description": "Некорректный JSON, ID вопроса или пустой текст",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Недействительный токен или пользователь не является автором квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Вопрос не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/answers/{id}": {
            "get": {
                "description": "Публичное представление ответа без признака is_correct.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Получить ответ",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID ответа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.PublicAnswerResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID ответа",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Ответ не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только автору квиза.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Удалить ответ",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID ответа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ответ удалён; тело ответа пустое"
                    },
                    "400": {
                        "description": "Некорректный ID ответа",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Недействительный токен или пользователь не является автором квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Ответ не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/answers/{id}/author": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает ответ с признаком is_correct. Доступно только автору квиза.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Получить ответ для автора",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID ответа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.AuthorAnswerResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID ответа",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Недействительный токен или пользователь не является автором квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Ответ не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "JPG/PNG до 5 MiB, до 20 мегапикселей. Поле multipart: file. Возвращает ID и временную ссылку S3.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Загрузить изображение вопроса",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Изображение",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/quiz.UploadedImage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/{id}": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Перенаправление на временную ссылку S3",
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "Ссылка S3"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "images"
                ],
                "summary": "Удалить неиспользуемое изображение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/{id}/url": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Получить свежую ссылку на изображение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz.UploadedImage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Текущий пользователь",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me/quizzes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Мои квизы с количеством вопросов и датой изменения",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/quiz.QuizSummary"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/questions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только автору квиза. Тело запроса должно содержать один JSON-объект без неизвестных полей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "Создать вопрос",
                "parameters": [
                    {
                        "description": "Новый вопрос",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quizhandler.CreateQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Вопрос создан; тело ответа пустое"
                    },
                    "400": {
                        "description": "Некорректный JSON, ID квиза или пустой текст",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Недействительный токен или пользователь не является автором квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Квиз не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/questions/{id}": {
            "get": {
                "description": "Возвращает вопрос. В текущей реализации поле answers равно null; ответы загружаются при получении квиза.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "Получить вопрос",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID вопроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.QuestionResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID вопроса",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Вопрос не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только автору квиза.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "Удалить вопрос",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID вопроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Вопрос удалён; тело ответа пустое"
                    },
                    "400": {
                        "description": "Некорректный ID вопроса",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Недействительный токен или пользователь не является автором квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Вопрос не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/questions/{id}/position": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только автору квиза. Позиция начинается с 1 и не может превышать число вопросов в квизе. Остальные вопросы сдвигаются. Тело запроса должно содержать один JSON-объект без неизвестных полей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "Изменить позицию вопроса",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID вопроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новая позиция",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quizhandler.ChangeQuestionPositionRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Позиция изменена; тело ответа пустое"
                    },
                    "400": {
                        "description": "Некорректный JSON, ID вопроса или позиция",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Недействительный токен или пользователь не является автором квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Вопрос не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quizzes": {
            "get": {
                "description": "Возвращает все квизы с вопросами и ответами без признака is_correct. Пагинации нет.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Получить список квизов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/quizhandler.QuizResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Автор определяется по JWT. Тело запроса должно содержать один JSON-объект без неизвестных полей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Создать квиз",
                "parameters": [
                    {
                        "description": "Новый квиз",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quizhandler.CreateQuizRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданный квиз",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.QuizResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON или пустое название",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Отсутствует или недействителен Bearer-токен",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quizzes/{id}/author": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Содержимое квиза для автора, включая правильные ответы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID квиза",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz.QuizContent"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "quiz.UploadedImage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "quizhandler.CreateAnswerRequest": {
            "type": "object",
            "required": [
                "question_id",
                "text_content"
            ],
            "properties": {
                "is_correct": {
                    "type": "boolean",
                    "default": false,
                    "example": true
                },
                "question_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "text_content": {
                    "description": "Текст не должен быть пустым после удаления пробелов по краям.",
                    "type": "string",
                    "minLength": 1,
                    "example": "Париж"
                }
            }
        },
...
        "respond.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите Bearer \u003cJWT\u003e. Токен можно получить через POST /auth/login с JSON {\"username\":\"alice\",\"password\":\"secret123\"}. Поле token уже содержит префикс Bearer. Регистрация: POST /auth/register с теми же полями.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}